+++ lotte/src/main/java/com/airbnb/lotte/animation/LotteAnimatableNumberValue.java
@@ -90,7 +90,7 @@ public class LotteAnimatableNumberValue implements LotteAnimatableValue<Number> 
             for (int i = 0; i < keyframes.length(); i++) {
                 JSONObject keyframe = keyframes.getJSONObject(i);
                 long frame = keyframe.getLong("t");
-                float timePercentage = (frame - startFrame) / durationFrames;
+                float timePercentage = (float) (frame - startFrame) / (float) durationFrames;
 
                 if (outValue != null) {
                     valueKeyframes.add(outValue);
